+++ tk205/__init__.py
@@ -9,3 +9,20 @@ def validate(file):
 
 def translate(input, output):
     dump(load(input),output)
+
+def translate_directory(source_dir, output_dir, clear=True):
+    output_extension = '.' + os.path.split(output_dir)[-1]
+    if clear:
+        clear_directory(output_dir)
+    for source in os.listdir(source_dir):
+        if '~$' not in source:  # Ignore temporary Excel files
+            source_path = os.path.join(source_dir,source)
+            base_name = os.path.basename(source_path)
+            file_name = os.path.splitext(base_name)[0]
+            output_path = os.path.join(output_dir,file_name + output_extension)
+            translate(source_path, output_path)
+
+def validate_directory(example_dir):
+    for example in os.listdir(example_dir):
+        if '~$' not in example:  # Ignore temporary Excel files
+            validate(os.path.join(example_dir,example))
